+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+# 1.1.0 (24 Nov 2014)
+* Fixed typo confusing `parent` with `collection` in meta
+* Routes for menus in theme locations now include complete tree with item order and nested children 
+* `description` attribute for individual items is now included in results
+
 # 1.0.0 (21 Jul 2014)
 First public release